+++ frontapp/src/routes/dashboard/+layout.server.js
@@ -1,0 +1,6 @@
+/** @type {import('./$types').LayoutServerLoad} */
+export function load({ locals }) {
+  return {
+    user: locals.user,
+  };
+}
