+++ frontapp/src/routes/dashboard/admin/attendance-management/+page.svelte
@@ -1,0 +1,1 @@
+<h1 class="text-2xl font-bold">出席登録</h1>
