+++ frontapp/src/routes/dashboard/admin/qr-code-reader/+page.svelte
@@ -1,0 +1,1 @@
+<h1 class="text-2xl font-bold">QRコード読み取り</h1>
