+++ frontapp/src/routes/dashboard/admin/user-management/+page.svelte
@@ -1,0 +1,1 @@
+<h1 class="text-2xl font-bold">ユーザー登録・管理</h1>
