+++ frontapp/src/routes/dashboard/admin/vorting-mvp/+page.svelte
@@ -1,0 +1,1 @@
+<h1 class="text-2xl font-bold">MVP投票</h1>
