+++ frontapp/src/routes/dashboard/root/notification/+page.svelte
@@ -1,0 +1,1 @@
+<h1 class="text-2xl font-bold">通知管理</h1>
