+++ frontapp/src/routes/dashboard/root/tournament-management/+page.svelte
@@ -1,0 +1,1 @@
+<h1 class="text-2xl font-bold">トーナメント生成・管理</h1>
