+++ frontapp/src/routes/dashboard/student/class-info/+page.svelte
@@ -1,0 +1,1 @@
+<h1 class="text-2xl font-bold">クラス情報</h1>
