+++ frontapp/src/routes/dashboard/student/issueqr-code/+page.svelte
@@ -1,0 +1,1 @@
+<h1 class="text-2xl font-bold">QRコード発行</h1>
