+++ frontapp/src/routes/dashboard/student/my-page/+page.svelte
@@ -1,0 +1,1 @@
+<h1 class="text-2xl font-bold">マイページ</h1>
