+++ frontapp/src/routes/dashboard/student/notification/+page.svelte
@@ -1,0 +1,1 @@
+<h1 class="text-2xl font-bold">通知</h1>
