+++ frontapp/src/routes/dashboard/student/score-list/+page.svelte
@@ -1,0 +1,1 @@
+<h1 class="text-2xl font-bold">点数一覧</h1>
